user情報取得処理をカスタムフック化
