+++ src/components/Authentication.tsx
@@ -4,6 +4,7 @@ import { Auth } from "@supabase/ui";
 import { AuthenticationPresenter } from "./AuthenticationPresenter";
 import { supabase } from "../../utils/supabaseClient";
 import { Loader } from "./Loader";
+import useFetchUser from "../hooks/useFetchUser";
 
 export type AuthenticationProps = {
   children: ReactNode;
@@ -11,28 +12,7 @@ export type AuthenticationProps = {
 
 export const Authentication: VFC<AuthenticationProps> = ({ children }) => {
   const { session } = Auth.useUser();
-  const [user, setUser] = useState<User | null>(null);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-  const [error, setError] = useState<ApiError | null>(null);
-
-  const getUserBySession = async (session: Session) => {
-    const { user, error } = await supabase.auth.api.getUser(
-      session.access_token
-    );
-    if (error) {
-      setError(error);
-      console.error(error.message, error.status);
-    }
-    if (user) setUser(user);
-  };
-
-  useEffect(() => {
-    if (!user && session) {
-      getUserBySession(session);
-    } else {
-      setIsLoading(false);
-    }
-  }, [user, session]);
+  const { user, error, isLoading } = useFetchUser(session);
 
   if (error) {
     return <Loader />; // TODO: should be replaced with `Error Boundary`
